resources/uhh-lt-event-classify/pyproject.toml
[tool.poetry]
name = "event-classification"
version = "0.1.0"
description = ""
license = "MIT"
authors = ["Hans Ole Hatzel <hhatzel@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
hydra-core = "1.2"
hydra-optuna-sweeper = "1.2"
conllu = "^4.4"
matplotlib = "3.5.2"
mlflow = "1.26.1"
numpy = "1.22.3"
omegaconf = "2.2.2"
pandas = "1.4.2"
scikit_learn = "1.1.1"
seaborn = "0.11.2"
spacy = "3.3.1"
statsmodels = "0.13.2"
thinc = "8.0.17"
torch = "1.11.0"
tqdm = "4.64.0"
transformers = "4.12.5"
typer = "0.4.1"
ujson = "5.3.0"
catma-gitlab = { git = "https://github.com/hatzel/catma_gitlab.git", rev = "54011e3f" }
scipy = "1.8.1"
de-dep-news-trf = {url = "https://github.com/explosion/spacy-models/releases/download/de_dep_news_trf-3.3.0/de_dep_news_trf-3.3.0-py3-none-any.whl"}
cupy = {version = "^10.5.0", optional = true}
torchserve = {version = "^0.6.0", optional = true}
torch-model-archiver = {version = "^0.6.0", optional = true}

[tool.poetry.extras]
gpu = ["cupy"]
torchserve = ["torchserve", "torch-model-archiver"]

[tool.poetry.dev-dependencies]
black = "^22.3"
isort = "^5.10"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
